bots/max/bot4_survey/requirements4.txt
max-sdk==1.0.0
requests==2.31.0
python-dotenv==1.0.0
gspread==5.11.0
google-auth==2.23.0